fix: 구독 기능 구현하며 발생한 서버/클라이언트 컴포넌트 에러 해결
