+++ src/app/(menu-header)/projects/[id]/page.tsx
@@ -1,9 +1,14 @@
+import { COMMENT_QUERY_KEY } from '@/constants';
+import { getComments } from '@/features/comment/services';
 import {
+  ProjectDetailContainer,
   ProjectDetailFallback,
-  ProjectDetailFetcher,
 } from '@/features/project/components';
 import { getProject } from '@/features/project/services';
+import { getQueryClient } from '@/libs/tanstack-query';
+import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
 import { Metadata } from 'next';
+import { notFound } from 'next/navigation';
 import { cache, Suspense } from 'react';
 
 type ProjectDetailPageProps = {
@@ -41,10 +46,29 @@ export default async function ProjectDetailPage({
 }: ProjectDetailPageProps) {
   const { id } = await params;
 
+  const project = await getProjectData(id);
+  const queryClient = getQueryClient();
+
+  await queryClient.prefetchInfiniteQuery({
+    queryKey: COMMENT_QUERY_KEY.COMMENTS(Number(id)),
+    queryFn: ({ pageParam }) =>
+      getComments(Number(id), pageParam.nextCursorTime, pageParam.nextCursorId),
+    staleTime: 1000 * 60,
+    initialPageParam: {
+      nextCursorId: 0,
+      nextCursorTime: null,
+    },
+  });
+
+  if (!project) {
+    notFound();
+  }
   return (
     <section className="pb-25">
       <Suspense fallback={<ProjectDetailFallback />}>
-        <ProjectDetailFetcher id={id} />
+        <HydrationBoundary state={dehydrate(queryClient)}>
+          <ProjectDetailContainer project={project} />
+        </HydrationBoundary>
       </Suspense>
     </section>
   );
