+++ src/features/project/components/index.ts
@@ -7,7 +7,6 @@ export * from './image-uploader';
 export * from './latest-projects';
 export * from './new-project-content';
 export * from './project-detail-container';
-export * from './project-detail-fetcher';
 export * from './project-gallery';
 export * from './projects-container';
 export * from './projects-fetcher';
