+++ src/features/project/components/project-detail-container/ProjectDetailContainer.tsx
@@ -1,7 +1,6 @@
 import { ChatbotButton } from '@/features/chatbot/components';
-import { cookies } from 'next/headers';
+import { Comments } from '@/features/comment/components';
 import Image from 'next/image';
-import { Comments } from '../../../comment/components/comments/Comments';
 import { ProjectDetail } from '../../schemas';
 import { Carousel } from '../carousel';
 import { Dashboard } from './Dashboard';
@@ -32,9 +31,6 @@ export async function ProjectDetailContainer({
     receivedPumatiCount,
     teamRank,
   } = project;
-
-  const cookieStore = await cookies();
-  const refreshToken = cookieStore.get('refreshToken')?.value;
 
   return (
     <div className="flex flex-col gap-1">
@@ -81,14 +77,12 @@ export async function ProjectDetailContainer({
       <div className="max-w-[25rem] w-full mx-auto">
         <Comments projectId={id} />
       </div>
-      {!!refreshToken && (
-        <ChatbotButton
-          title={title}
-          term={term}
-          teamNumber={teamNumber}
-          projectId={id.toString()}
-        />
-      )}
+      <ChatbotButton
+        title={title}
+        term={term}
+        teamNumber={teamNumber}
+        projectId={id.toString()}
+      />
     </div>
   );
 }
