--- src/features/project/components/project-detail-fetcher/index.ts
@@ -1,1 +1,0 @@
-export * from './ProjectDetailFetcher';
