+++ src/features/subscribe/hooks/index.ts
@@ -1,2 +1,3 @@
+export * from './useCheckSubscriptionState';
 export * from './useSubscribe';
 export * from './useUnSubscribe';
