+++ src/hooks/useOutsideClick.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { RefObject, useEffect } from 'react';
 
 /**
